+++ migrations/2024-09-01-094614_remove-preferred-merchant-from-users/up.sql
@@ -1,0 +1,2 @@
+-- Your SQL goes here
+ALTER TABLE users DROP COLUMN preferred_merchant_id;